Save feeAmount in pennies

+++ src/main/java/uk/gov/hmcts/reform/unspec/model/ClaimFee.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @Builder
@@ -13,6 +13,6 @@ public class ClaimFee {
     private String code;
     private String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING)
-    private BigDecimal feeAmount;
+    private BigInteger feeAmount;
     private String version;
 }
